# api_rede_social/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Protótipo de Rede Social",
        "description": "API protótipo de rede social.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Listar todas as tags.",
                "operationId": "e9df4702b5217bc5d1b56361b7d9ef68",
                "responses": {
                    "200": {
                        "description": "Lista de tags retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID da tag",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "description": "Nome da tag",
                                                "type": "string",
                                                "example": "Tecnologia"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Cadastrar uma nova tag.",
                "operationId": "5cfcf67f1f41d16305affab21cadfc99",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome da tag",
                                        "type": "string",
                                        "example": "Tecnologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID da tag",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "description": "Nome da tag",
                                            "type": "string",
                                            "example": "Tecnologia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou faltando parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar a tag."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tags/{id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Exibir os dados de uma tag específica.",
                "operationId": "73e65c8d6ae972fe586a6eda262458e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tag",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag encontrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID da tag",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "description": "Nome da tag",
                                            "type": "string",
                                            "example": "Tecnologia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tag não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Tag não encontrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tags"
                ],
                "summary": "Atualizar os dados de uma tag específica.",
                "operationId": "6143ac142a8bb3497893ccff88fb7ce7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tag",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome da tag",
                                        "type": "string",
                                        "example": "Tecnologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tag atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID da tag",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "description": "Nome da tag",
                                            "type": "string",
                                            "example": "Tecnologia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou faltando parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao atualizar a tag."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Deletar uma tag específica.",
                "operationId": "3b04880727363a917f1c9ea1b729e6de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tag",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tag deletada com sucesso"
                    },
                    "404": {
                        "description": "Tag não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Tag não encontrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}